refactor(list-group-item): rename typings to types
