+++ packages/list-group-item/index.d.ts
@@ -1,5 +1,2 @@
-import ListGroupItem from './typings/ListGroupItem';
-
-export { default as ListGroupItemStatus } from './typings/ListGroupItemStatus';
-
-export default ListGroupItem;
+export { default } from './types/ListGroupItem';
+export { default as ListGroupItemStatus } from './types/ListGroupItemStatus';
