+++ packages/list-group-item/index.js
@@ -1,5 +1,2 @@
-import ListGroupItem from './src/ListGroupItem';
-
+export { default } from './src/ListGroupItem';
 export { default as ListGroupItemStatus } from './src/ListGroupItemStatus';
-
-export default ListGroupItem;
